refactor(xslate): extract expr/value/lib modules + emit-context seam
Mirror the Mojo adapter's domain-module decomposition (issue #2018 track
D) for the Text::Xslate (Kolon) adapter. Pure free functions and the
ParsedExpr emitters move out behind a narrow `XslateEmitContext` seam;
the main adapter keeps its recursive lowering but delegates the extracted
clusters. Behaviour is unchanged — adapter-tests conformance stays
byte-identical (full suite 353 pass / 0 fail).

New layout under packages/adapter-xslate/src/adapter/:
  emit-context.ts        XslateEmitContext interface (the seam)
  lib/types.ts           PrimitiveSpec, XslateRenderCtx, XslateAdapterOptions
  lib/constants.ts       XSLATE_TEMPLATE_PRIMITIVES + emit map
  lib/kolon-naming.ts    escapeKolonSingleQuoted, kolonHashKey
  lib/ir-scope.ts        resolveJsxChildrenProp, collectRootScopeNodes,
                         referencedVarsAreAvailable
  value/parsed-literal.ts parsePureStringLiteral + string-type helpers
  expr/operand.ts        isStringTypedOperand (parallel of Mojo; see note)
  expr/array-method.ts   renderArrayMethod/Sort/Reduce/Flat/FlatMap
  expr/emitters.ts       XslateFilterEmitter, XslateTopLevelEmitter
